pyproject.toml
[project]
name = "agentram"
version = "0.1.0"
description = "Persistent memory layer for AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "mcp>=1.0.0",
    "gitpython>=3.1.40",
    "typer>=0.12.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
semantic = [
    "numpy>=1.26.0",
    "sentence-transformers>=2.5.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
ram-server = "agentram.cli:server"
ram-cli = "agentram.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentram"]
